requirements.txt
flask>=2.0
tensorflow>=2.11
pillow
numpy
gunicorn
requests
